doing some refactoring, fixing some bugs in origin/master!
