+++ src/routes/routes.js
@@ -1,6 +1,6 @@
-const tweetRoutes = require('./tweet-routes');
-const authRoutes = require('./auth-routes');
-const tarinRoutes = require('./tarin-routes');
+const tweetRoutes = require('./tweet.routes');
+const authRoutes = require('./auth.route');
+const tarinRoutes = require('./tarin.route');
 const userRoutes = require('./user.route');
 
 module.exports = (app) => {
